Separate form logic from form view
